backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pandas
python-dotenv
openpyxl
python-multipart